+++ src/lib/data/near.js
@@ -208,6 +208,7 @@ async function _initNear({
   keyStore,
   selector,
   walletConnectCallback = () => {},
+  customElements = {},
 }) {
   if (!config) {
     config = {};
@@ -224,6 +225,11 @@ async function _initNear({
     config = Object.assign({}, TestNearConfig, config);
   }
   config.walletConnectCallback = walletConnectCallback;
+  config.customElements = Object.assign(
+    {},
+    config.customElements,
+    customElements
+  );
   keyStore = keyStore ?? new nearAPI.keyStores.BrowserLocalStorageKeyStore();
 
   const nearConnection = await nearAPI.connect(