+++ webpack.config.js
@@ -33,12 +33,6 @@ module.exports = function (env) {
             amd: "react-dom",
             root: "ReactDOM",
           },
-          "react-router-dom": {
-            commonjs: "react-router-dom",
-            commonjs2: "react-router-dom",
-            amd: "react-router-dom",
-            root: "ReactRouterDOM",
-          },
         },
       ],
       module: {